+++ Frends.Oracle.ExecuteQuery/Frends.Oracle.ExecuteQuery.Tests/UnitTests.cs
@@ -1,9 +1,10 @@
 using NUnit.Framework;
 using System;
 using System.Threading;
-using System.Collections.Generic;
 using Frends.Oracle.ExecuteQuery.Definitions;
 using Newtonsoft.Json.Linq;
+using Oracle.ManagedDataAccess.Client;
+using System.Threading.Tasks;
 
 namespace Frends.Oracle.ExecuteQuery.Tests
 {
@@ -11,292 +12,252 @@ namespace Frends.Oracle.ExecuteQuery.Tests
     /// Oracle test database is needed to run these test. 
     /// </summary>
     [TestFixture]
-    [Ignore("Tests needs new Ubuntu based workflow")]
     class TestClass
     {
         /// <summary>
         /// Connection string for Oracle database.
         /// </summary>
-        private static string _connectionString = Environment.GetEnvironmentVariable("HiQ_OracleDb_connectionString");
-        private static int _connectionTimeout = 30;
+        private readonly static string _schema = "test_user";
+        private readonly static string _connectionString = $"Data Source = (DESCRIPTION = (ADDRESS = (PROTOCOL = TCP)(HOST = localhost)(PORT = 51521))(CONNECT_DATA = (SERVICE_NAME = XEPDB1))); User Id = {_schema}; Password={_schema};";
+        private readonly static string _connectionStringSys = "Data Source = (DESCRIPTION = (ADDRESS = (PROTOCOL = TCP)(HOST = localhost)(PORT = 51521))(CONNECT_DATA = (SERVICE_NAME = XEPDB1))); User Id = sys; Password=mysecurepassword; DBA PRIVILEGE=SYSDBA";
 
         /// <summary>
         /// Global variables.
         /// </summary>
-        private static ConnectionProperties _connectionProperties;
-        private static QueryProperties _queryProperties;
+        private static Input _input;
         private static Options _options;
 
+        #region OneTimeSetup&TearDown
         [OneTimeSetUp]
-        public void SetUp()
+        public void OneTimeSetUp()
         {
-            _connectionProperties = new ConnectionProperties { ConnectionString = _connectionString, TimeoutSeconds = _connectionTimeout };
-            _queryProperties = new QueryProperties
-            {
-                Query = "create table workers(" +
-                "id NUMBER, " +
-                "first_name varchar2(50), " +
-                "last_name varchar2(50), " +
-                "start_date date, " +
-                "primary key(id))"
-            };
-            _options = new Options { ThrowErrorOnFailure = true, BindParameterByName = true };
+            Helpers.TestConnectionBeforeRunningTests(_connectionStringSys);
 
-            Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            using var con = new OracleConnection(_connectionStringSys);
+            con.Open();
+            Helpers.CreateTestUser(con);
+            con.Close();
+            if (con.State == System.Data.ConnectionState.Open)
+                con.Close();
         }
 
         [OneTimeTearDown]
+        public void OneTimeTearDown()
+        {
+            using var con = new OracleConnection(_connectionStringSys);
+            con.Open();
+            Helpers.DropTestUser(con);
+            con.Close();
+            if (con.State == System.Data.ConnectionState.Open)
+                con.Close();
+        }
+        #endregion OneTimeSetup&TearDown
+        #region Setup&TearDown
+        [SetUp]
+        public void Setup()
+        {
+            _input = new Input
+            {
+                ConnectionString = _connectionString,
+            };
+
+            _options = new Options
+            {
+                ThrowErrorOnFailure = true,
+                BindParameterByName = true,
+                TimeoutSeconds = 30
+            };
+
+            using var con = new OracleConnection(_connectionStringSys);
+            con.Open();
+            Helpers.CreateTestTable(con);
+            con.Close();
+            if (con.State == System.Data.ConnectionState.Open)
+                con.Close();
+        }
+
+        [TearDown]
         public void TearDown()
         {
-            _connectionProperties = new ConnectionProperties { ConnectionString = _connectionString, TimeoutSeconds = _connectionTimeout };
-            _queryProperties = new QueryProperties
-            {
-                Query = "drop table workers"
-            };
-            _options = new Options { ThrowErrorOnFailure = true };
-
-            Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());        
+            using var con = new OracleConnection(_connectionStringSys);
+            con.Open();
+            Helpers.DropTestTable(con);
+            con.Close();
+            if (con.State == System.Data.ConnectionState.Open)
+                con.Close();
         }
+        #endregion Setup&TearDown
 
         [Test]
-        public void ExecuteQuery_InsertWithParameters()
+        public async Task ExecuteQuery_InsertWithParameters()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values (:id, :name, 'Meikäläinen')";
+            _input.Parameters = new QueryParameter[]
             {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values (:id, :name, 'Meikäläinen')",
-                Parameters = new QueryParameter[]
-                {
-                    new QueryParameter { Name = "name", Value = "Matti", DataType = QueryParameterType.Varchar2 },
-                    new QueryParameter { Name = "id", Value = 3, DataType = QueryParameterType.Int32 },
-                }
+                new QueryParameter { Name = "name", Value = "Matti", DataType = QueryParameterType.Varchar2 },
+                new QueryParameter { Name = "id", Value = 3, DataType = QueryParameterType.Int32 },
             };
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.That(result, Is.Not.Null);
-            Assert.AreEqual(1, result.RowsAffected);
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "select first_name from workers where id = 3"
-            };
+            _input.Query = "select first_name from workers where id = 3";
 
-            var expected = "Matti";
-            result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
-            result.Output.ForEach(i => Console.WriteLine("{0}\t", i));
-            Assert.AreEqual(result.Output.GetType(), typeof(List<JObject>));
-            Assert.AreEqual(expected, result.Output[0]["FIRST_NAME"].ToString());
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual(typeof(JArray), result.Output.GetType());
+            Assert.AreEqual("Matti", (string)result.Output[0]["FIRST_NAME"]);
         }
 
         [Test]
-        public void ExecuteQuery_WithAllValues()
+        public async Task ExecuteQuery_WithAllValues()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert " +
-                "into workers values (1, 'Matti', 'Meikäläinen', DATE '2022-04-12')"
-            };
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            _input.Query = "insert " +
+                "into workers values (1, 'Matti', 'Meikäläinen', DATE '2022-04-12')";
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.IsNotNull(result);
             Assert.AreEqual(true, result.Success);
-            Assert.AreEqual(1, result.RowsAffected);
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
         }
 
         [Test]
-        public void ExecuteQuery_InsertMultipleRowsIntoTable()
+        public async Task ExecuteQuery_InsertMultipleRowsIntoTable()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert all " +
+            _input.Query = "insert all " +
                 "into workers (id, first_name, last_name) values (1, 'Matti', 'Meikäläinen') " +
                 "into workers (id, first_name, last_name) values (2, 'Teppo', 'Teikäläinen') " +
-                "select * from dual",
-            };
+                "select * from dual";
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.IsNotNull(result);
             Assert.AreEqual(true, result.Success);
-            Assert.AreEqual(2, result.RowsAffected);
+            Assert.AreEqual(2, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
+            _input.Query = "select * from workers " +
+                "where id = :id";
+            _input.Parameters = new QueryParameter[]
             {
-                Query = "select * from workers " +
-                "where id = :id",
-                Parameters = new QueryParameter[]
-                {
-                    new QueryParameter { Name = "id", Value = 1, DataType = QueryParameterType.Int32 }
-                }
+                new QueryParameter { Name = "id", Value = 1, DataType = QueryParameterType.Int32 }
             };
 
-            result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.That(result, Is.Not.Null);
             Assert.AreEqual(true, result.Success);
-            Assert.AreEqual(0, result.RowsAffected);
+            Assert.AreEqual("Matti", (string)result.Output[0]["FIRST_NAME"]);
+            Assert.AreEqual("Meikäläinen", (string)result.Output[0]["LAST_NAME"]);
         }
 
         [Test]
-        public void ExecuteQuery_InsertMultipleRowsWithMultipleParameters()
+        public async Task ExecuteQuery_InsertMultipleRowsWithMultipleParameters()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert all " +
+            _input.Query = "insert all " +
                 "into workers (id, first_name, last_name) values (:id1, :fname1, :lname1) " +
                 "into workers (id, first_name, last_name) values (:id2, :fname2, :lname2) " +
-                "select * from dual",
-                Parameters = new QueryParameter[]
-                {
-                    new QueryParameter { Name = "id1", Value = 1, DataType = QueryParameterType.Int32 },
-                    new QueryParameter { Name = "fname1", Value = "Matti", DataType = QueryParameterType.Varchar2 },
-                    new QueryParameter { Name = "lname1", Value ="Meikäläinen", DataType = QueryParameterType.Varchar2 },
-                    new QueryParameter { Name = "id2", Value = 2, DataType = QueryParameterType.Int32 },
-                    new QueryParameter { Name = "fname2", Value = "Teppo", DataType = QueryParameterType.Varchar2 },
-                    new QueryParameter { Name = "lname2", Value = "Teikäläinen", DataType = QueryParameterType.Varchar2 }
-                }
+                "select * from dual";
+            _input.Parameters = new QueryParameter[]
+            {
+                new QueryParameter { Name = "id1", Value = 1, DataType = QueryParameterType.Int32 },
+                new QueryParameter { Name = "fname1", Value = "Matti", DataType = QueryParameterType.Varchar2 },
+                new QueryParameter { Name = "lname1", Value ="Meikäläinen", DataType = QueryParameterType.Varchar2 },
+                new QueryParameter { Name = "id2", Value = 2, DataType = QueryParameterType.Int32 },
+                new QueryParameter { Name = "fname2", Value = "Teppo", DataType = QueryParameterType.Varchar2 },
+                new QueryParameter { Name = "lname2", Value = "Teikäläinen", DataType = QueryParameterType.Varchar2 }
             };
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.IsNotNull(result);
             Assert.AreEqual(true, result.Success);
-            Assert.AreEqual(2, result.RowsAffected);
+            Assert.AreEqual(2, (int)result.Output["AffectedRows"]);
         }
 
         [Test]
-        public void ExecuteQuery_Update()
+        public async Task ExecuteQuery_Update()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values (1, 'Matti', 'Meikäläinen')"
-            };
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values (1, 'Matti', 'Meikäläinen')";
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
-            Assert.IsNotNull(result);
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "update workers " +
+            _input.Query = "update workers " +
                 "set first_name = 'Saija', " +
                 "last_name = 'Saijalainen' " +
-                "where id = 1"
-            };
+                "where id = 1";
 
-            result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
-            Assert.AreEqual(1, result.RowsAffected);
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "select first_name from workers where id = 1"
-            };
+            _input.Query = "select first_name from workers where id = 1";
 
-            var expected = "Saija";
-            result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
-            Assert.AreEqual(expected, result.Output[0]["FIRST_NAME"].ToString());
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual("Saija", (string)result.Output[0]["FIRST_NAME"]);
         }
 
         [Test]
-        public void ExecuteQuery_SelectWithNonExistingRow()
+        public async Task ExecuteQuery_SelectWithNonExistingRow()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values (1, 'Matti', 'Meikäläinen')"
-            };
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values (1, 'Matti', 'Meikäläinen')";
 
-            Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "select first_name from workers where id = 2"
-            };
+            _input.Query = "select first_name from workers where id = 2";
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.That(result, Is.Not.Null);
-            Assert.That(result.Output, Is.Not.Null);
+            Assert.AreEqual("[]", result.Output.ToString());
             Assert.That(result.Success, Is.True);
-            Assert.AreEqual(0, result.Output.Count);
-            Assert.AreEqual(0, result.RowsAffected);
         }
 
         [Test]
-        public void ExecuteQuery_WithoutThrowErrorOnFailure()
+        public async Task ExecuteQuery_WithoutThrowErrorOnFailure()
         {
-            ClearTable();
-            _options = new Options { ThrowErrorOnFailure = false };
+            _options.ThrowErrorOnFailure = false;
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values ('Matti', 1, 'Meikäläinen')",
-            };
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values ('Matti', 1, 'Meikäläinen')";
 
-            var result = (ErrorResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.That(result, Is.Not.Null);
             Assert.AreEqual(false, result.Success);
-            Assert.AreEqual("ORA-01722: invalid number", result.Error);
+            Assert.AreEqual("ORA-01722: invalid number", result.Message);
         }
 
         [Test]
         public void ExecuteQuery_ThatThrowsException()
         {
-            ClearTable();
-            _queryProperties = new QueryProperties
-            {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values ('Matti', 1, 'Meikäläinen')",
-            };
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values ('Matti', 1, 'Meikäläinen')";
 
-            var error = Assert.Throws<Exception>(() => Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken()));
+            var error = Assert.ThrowsAsync<Exception>(async () => await Oracle.ExecuteQuery(_input, _options, new CancellationToken()));
             Assert.AreEqual("ORA-01722: invalid number", error.Message);
         }
 
         [Test]
-        public void ExecuteQuery_InsertWithBindParameterByNameFalse()
+        public async Task ExecuteQuery_InsertWithBindParameterByNameFalse()
         {
-            ClearTable();
             _options.BindParameterByName = false;
-            _queryProperties = new QueryProperties
+            _input.Query = "insert " +
+                "into workers (id, first_name, last_name) values (:p, :p, :p)";
+            _input.Parameters = new QueryParameter[]
             {
-                Query = "insert " +
-                "into workers (id, first_name, last_name) values (:p, :p, :p)",
-                Parameters = new QueryParameter[]
-                {
-                    new QueryParameter { Name = "p", Value = 1, DataType = QueryParameterType.Int32 },
-                    new QueryParameter { Name = "p", Value = "Matti", DataType = QueryParameterType.Varchar2 },
-                    new QueryParameter { Name = "p", Value ="Meikäläinen", DataType = QueryParameterType.Varchar2 },
-                }
+                new QueryParameter { Name = "p", Value = 1, DataType = QueryParameterType.Int32 },
+                new QueryParameter { Name = "p", Value = "Matti", DataType = QueryParameterType.Varchar2 },
+                new QueryParameter { Name = "p", Value ="Meikäläinen", DataType = QueryParameterType.Varchar2 },
             };
 
-            var result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            var result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
             Assert.That(result, Is.Not.Null);
             Assert.AreEqual(true, result.Success);
+            Assert.AreEqual(1, (int)result.Output["AffectedRows"]);
 
-            _queryProperties = new QueryProperties
-            {
-                Query = "select first_name, last_name from workers where id = 1"
-            };
+            _input.Query = "select first_name, last_name from workers where id = 1";
 
-            result = (QueryResult)Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
-            Assert.AreEqual("Matti", result.Output[0]["FIRST_NAME"].ToString());
-            Assert.AreEqual("Meikäläinen", result.Output[0]["LAST_NAME"].ToString());
-        }
-
-        /// <summary>
-        /// Helper method to clear Oracle database table workers
-        /// </summary>
-        public void ClearTable()
-        {
-            _options.BindParameterByName = true;
-            _queryProperties = new QueryProperties
-            {
-                Query = "delete from workers"
-            };
-            Oracle.ExecuteQuery(_connectionProperties, _queryProperties, _options, new CancellationToken());
+            result = await Oracle.ExecuteQuery(_input, _options, new CancellationToken());
+            Assert.AreEqual("Matti", (string)result.Output[0]["FIRST_NAME"]);
+            Assert.AreEqual("Meikäläinen", (string)result.Output[0]["LAST_NAME"]);
         }
     }
 }