+++ Frends.Oracle.ExecuteQuery/Frends.Oracle.ExecuteQuery/Definitions/QueryParameterType.cs
@@ -1,0 +1,40 @@
+﻿namespace Frends.Oracle.ExecuteQuery.Definitions;
+
+/// <summary>
+/// Enumerator representing oracle parameter data types
+/// </summary>
+#pragma warning disable 1591 // Self-explanatory.
+public enum QueryParameterType
+{
+    NVarchar2,
+    Varchar2,
+    NChar,
+    Char,
+    Int16,
+    Int32,
+    Int64,
+    Double,
+    Decimal,
+    Long,
+    LongRaw,
+    Boolean,
+    Date,
+    TimeStamp,
+    TimeStampLTZ,
+    TimeStampTZ,
+    XmlType,
+    Raw,
+    BFile,
+    BinaryDouble,
+    BinaryFloat,
+    Blob,
+    Byte,
+    Clob,
+    NClob,
+    IntervalDS,
+    IntervalYM,
+    RefCursor,
+    Single
+
+}
+