+++ Frends.Oracle.ExecuteQuery/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.1.0] - 2022-10-26
+### Fixed
+- [Breaking] Simplified implementation and merged ConnectionProperties and QueryProperties classes into one Input class.
+- Changed the implementation to resemble other database related tasks.
+- Modified the task to be asynchronous.
+- Changed the task use different methods when doing 'select' query and when doing other queries.
+- Updated workflow for Linux based testing and enabled the tests.
+
 ## [1.0.0] - 2022-04-22
 ### Added
 - Initial implementation.